+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - New `http` & `file` features which give a way to disable resolving external stylesheets and reduce the compiled artifacts size.
 
+### Fixed
+
+- `!important` rules not overriding inlined styles. [#152](https://github.com/Stranger6667/css-inline/issues/152)
+
 ## [0.8.1] - 2022-04-01
 
 ### Fixed